Move cmuxSymbolPixelSize to a shared file; reuse in SidebarHelpMenuButton
Addresses review feedback on the helper's discoverability: `cmuxSymbolPixelSize`
lived in `BrowserPanelView.swift` (module-visible but siloed) while
`SidebarHelpMenuButton` re-implemented the same four-step
`resizable/scaledToFit/fontWeight/frame` pattern inline.

Move the helper to the shared `RenderableSystemSymbol.swift` and have both the
browser toolbar and `SidebarHelpMenuButton` use it, removing the duplicated
inline pattern. No behavior change.
